Move test_cat_patches to test_patch.py
